requirements.txt
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.0
django-filter==2.3.0
django-mathfilters==1.0.0
django-modelcluster==5.0.2
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
ipython==7.17.0
ipython-genutils==0.2.0
isort==5.4.2
jdcal==1.4.1
jedi==0.17.2
l18n==2018.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
prompt-toolkit==3.0.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.10
wagtail==2.10
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
